Partial restructuring of the repo
