+++ src/api/utils/index.ts
@@ -1,0 +1,49 @@
+import type { IncludedQueryParams } from "../interfaces";
+
+export const getTransferQueryParams = (): IncludedQueryParams => {
+  return {
+    resource: {
+      type: true,
+      id: true,
+    },
+    toDomain: {
+      name: true,
+      lastIndexedBlock: true,
+      id: true,
+    },
+    fromDomain: {
+      name: true,
+      lastIndexedBlock: true,
+      id: true,
+    },
+    fee: {
+      id: true,
+      amount: true,
+      tokenAddress: true,
+      tokenSymbol: true,
+      decimals: true,
+    },
+    deposit: {
+      txHash: true,
+      blockNumber: true,
+      depositData: true,
+      handlerResponse: true,
+      timestamp: true,
+    },
+    execution: {
+      txHash: true,
+      blockNumber: true,
+      timestamp: true,
+    },
+    account: {
+      id: true,
+      addressStatus: true,
+    },
+  };
+};
+
+export class NotFound extends Error {
+  constructor(message: string) {
+    super(message);
+  }
+}
